move custom hooks to "hooks" folder
