+++ src/components/LoadingPage/index.js
@@ -1,2 +1,1 @@
 export { default } from './LoadingPage';
-export { default as useLoadingPage } from './useLoadingPage';
