+++ src/components/SignupModal/SignupModal.jsx
@@ -2,8 +2,8 @@ import React, { useState, useContext } from 'react';
 import styled from 'styled-components';
 import { useHistory } from 'react-router-dom';
 import { AuthContext } from '../../auth/Auth';
-import { api, extractTokenFromResponse, extractInfoFromToken } from './../../apis';
-import useForm from '../../pages/OfferModal/ProfilePage/SubPages/useForm';
+import { api, extractTokenFromResponse, extractInfoFromToken } from '../../apis';
+import useForm from '../../hooks/useForm';
 import Modal from '../Modal';
 import Button from '../Button';
 import Input from '../Input';
