+++ src/components/ToggleContent/index.js
@@ -1,3 +1,2 @@
 export { default } from './ToggleContent';
-export { default as useToggleContent } from './useToggleContent';
 export { default as withToggleContent } from './withToggleContent';
