+++ src/hooks/useCollapsible/Collapsible.module.scss
