+++ src/hooks/useCollapsible/useCollapsible.jsx
@@ -3,7 +3,7 @@ import className from 'classnames/bind';
 
 import styles from './Collapsible.module.scss';
 
-import Button from '../../../../../components/Button';
+import Button from '../../components/Button';
 
 const cx = className.bind(styles);
